fix arrow and sfx when changing submenu
BaseMenu now contains a property "is_active" that is set to true only
when set visible after grabing focus on the default button

FocusArrow is now teleported when Menu is inactive (on default focus)

Sound FX when changing focus is now handled by BaseMenu, for all buttons
part of a "focusable" group. No sound is played if menu is not active.